+++ custom_components/unisenza_plus/__init__.py
@@ -1,4 +1,5 @@
 """The Unisenza Plus integration."""
+
 from __future__ import annotations
 
 from homeassistant.config_entries import ConfigEntry